flashdrv: build as a static lib + test program
